refactor: 도메인 폴더 분리 및 변경점 수정
