+++ tests/services/test_comment_service.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from datetime import datetime
 
-from app.schemas.comment_create_dto import CommentCreateDTO
+from app.domains.comment import Comment
 from app.services.comment_service import update_comment
 
 
@@ -14,7 +14,7 @@ class Test(TestCase):
         content = "content"
         parent_id = "0"
         comment_date = datetime.now()
-        comment = CommentCreateDTO(id=id, user_id=user_id, content=content, parent_id=parent_id, comment_date=comment_date)
+        comment = Comment(id=id, user_id=user_id, content=content, parent_id=parent_id, comment_date=comment_date)
 
         # When
         updated_comment = update_comment(comment)
